requirements.txt
Django==1.4.5
django-cms==2.3.5
